Make ephemeral test reruns include temporary variant generation
Since the default assumption on avocado is that the variant name of
each test is unique, it creates and expects unique folder names for
each result folder. We should respect that assumption by adding a
randomly generated variant instead of patching the folder collision
thus allowing us to match the displayed test name with the correct
result folder. The temporary variant is generated for each ephemeral
test on each run thus reflecting the fact that these tests are fully
rerunnable with the sole purpose of obtaining a given setup.