Generalize from vm unique to host aware parameters
Any vm unique parameters can be generated with the suffix operator
and producing vm unique values is up to the configuration side which
should not be complicated further with code-only postprocessing stage.

We still have to get rid of the additional post-processing due to
host/container shared functionality like LVM and general storage
as well as network interfaces.