Merge parameter validation and printing for more performance
With the cost of a bit more method-bound complexity we can skip
another parser retrieval and thus improve parsing speed by another
third. It is now a bit faster than the pre-refactoring parsing.

It can be even faster and handle drastically more tests if the
cartesian config side is improved and in particular the behavior
of the recursive dictionary generator after join operation.